+++ CHANGELOG.md
@@ -4,7 +4,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-(Nothing yet)
+### Added
+- jsdoc documentation for some methods
+- blocking task running if task belongs to instance of queue that is destroyed
 
 ## [1.0.2] 2018-11-25
 ### Fixed